chore(ssi): refactor webhook code to its own file
This commit refactors the webhook code to its own file. This is the
first step to refactoring the tests and pulling them up to test
injection at the webhook level.
