+++ pkg/clusteragent/admission/mutate/autoinstrumentation/config.go
@@ -127,22 +127,6 @@ func NewConfig(datadogConfig config.Component) (*Config, error) {
 	}, nil
 }
 
-// WebhookConfig use to store options from the config.Component for the autoinstrumentation webhook
-type WebhookConfig struct {
-	// IsEnabled is the flag to enable the autoinstrumentation webhook.
-	IsEnabled bool
-	// Endpoint is the endpoint to use for the autoinstrumentation webhook.
-	Endpoint string
-}
-
-// NewWebhookConfig retrieves the configuration for the autoinstrumentation webhook from the datadog config
-func NewWebhookConfig(datadogConfig config.Component) *WebhookConfig {
-	return &WebhookConfig{
-		IsEnabled: datadogConfig.GetBool("admission_controller.auto_instrumentation.enabled"),
-		Endpoint:  datadogConfig.GetString("admission_controller.auto_instrumentation.endpoint"),
-	}
-}
-
 // LanguageDetectionConfig is a struct to store the configuration for the language detection. It can be populated using
 // the datadog config through NewLanguageDetectionConfig.
 type LanguageDetectionConfig struct {
